+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from './constants'
+export * from './location-hash'
 export * from './unforgettable-sdk'
 export * from './utils'